Meta_Collector: introduce lazy sorting
In case only very few elements are needed from current_objects,
sorting the vector and using binary search is quite expensive. In this
case, a linear search would be cheaper.

To avoid much overhead when requesting many elements, we use a small
threshold value of 10 to switch from linear to sorting / binary search.